refactor(game/ui): change directory
