+++ service/app/src/app/_components/GameSection.tsx
@@ -8,7 +8,7 @@ import { GameListItem } from "@/entities/game"
 import { getDefaultGame } from "@/entities/game/api/getDefaultGame"
 import { getGameDetail } from "@/entities/game/api/getGameDetail"
 import { useGameEntryNavigation } from "@/entities/game/hooks/useGameEntryNavigation"
-import { GamePreview } from "@/entities/game/ui/components/gamePreview"
+import { GamePreview } from "@/entities/game/ui/gamePreview"
 import * as GameCard from "@/shared/gameCard"
 import SSRSafeSuspense from "@/shared/SSRSafeSuspense"
 interface GameSectionProps {
