+++ service/app/src/app/dashboard/components/dashboardGameSection.tsx
@@ -5,8 +5,8 @@ import { useIntersectionObserver } from "react-simplikit"
 
 import { useGamePreview } from "@/entities/game/hooks/useGamePreview"
 import { useInfiniteMyGames } from "@/entities/game/model/useInfiniteMyGames"
-import { GameCardOptions } from "@/entities/game/ui/components/gameCardOptions"
-import { GameLibrarySkeleton } from "@/entities/game/ui/components/gameLibrarySkeleton"
+import { GameCardOptions } from "@/entities/game/ui/gameCardOptions"
+import { GameLibrarySkeleton } from "@/entities/game/ui/gameLibrarySkeleton"
 import * as GameCard from "@/shared/gameCard"
 
 import { useDashboardGameActions } from "../hooks/useDashboardGameActions"
