+++ service/app/src/app/games/components/gamesLibrarySection.tsx
@@ -5,7 +5,7 @@ import { useIntersectionObserver } from "react-simplikit"
 
 import { useGamePreview } from "@/entities/game/hooks/useGamePreview"
 import { useInfiniteGameList } from "@/entities/game/model/useInfiniteGameList"
-import { GameLibrarySkeleton } from "@/entities/game/ui/components/gameLibrarySkeleton"
+import { GameLibrarySkeleton } from "@/entities/game/ui/gameLibrarySkeleton"
 import * as GameCard from "@/shared/gameCard"
 
 import { filterInput } from "../utils/filterInput"
