--- service/app/src/entities/game/ui/components/index.ts
@@ -1,1 +1,0 @@
-export { GameLibraryGrid } from "./gameLibraryGrid"
