+++ service/app/src/entities/game/ui/gameCardOptions.tsx
