+++ service/app/src/entities/game/ui/gamePreview.tsx
